+++ apps/ios/ADE/Views/Lanes/LaneComponents.swift
@@ -1,0 +1,445 @@
+import SwiftUI
+
+// MARK: - Glass section
+
+struct GlassSection<Content: View>: View {
+  let title: String
+  let subtitle: String?
+  let content: Content
+
+  init(title: String, subtitle: String? = nil, @ViewBuilder content: () -> Content) {
+    self.title = title
+    self.subtitle = subtitle
+    self.content = content()
+  }
+
+  var body: some View {
+    VStack(alignment: .leading, spacing: 12) {
+      VStack(alignment: .leading, spacing: 3) {
+        Text(title)
+          .font(.subheadline.weight(.semibold))
+          .foregroundStyle(ADEColor.textPrimary)
+        if let subtitle {
+          Text(subtitle)
+            .font(.caption)
+            .foregroundStyle(ADEColor.textSecondary)
+        }
+      }
+      content
+    }
+    .adeGlassCard(cornerRadius: 16, padding: 14)
+  }
+}
+
+// MARK: - Lane status indicator
+
+struct LaneStatusIndicator: View {
+  @Environment(\.accessibilityReduceMotion) private var reduceMotion
+
+  let bucket: String
+  var size: CGFloat = 10
+
+  @State private var isPulsing = false
+
+  var body: some View {
+    Circle()
+      .fill(runtimeTint(bucket: bucket))
+      .frame(width: size, height: size)
+      .shadow(color: runtimeTint(bucket: bucket).opacity(isAnimating ? 0.5 : 0), radius: isAnimating ? 6 : 0)
+      .scaleEffect(isPulsing && isAnimating ? 1.3 : 1.0)
+      .animation(ADEMotion.pulse(reduceMotion: reduceMotion), value: isPulsing)
+      .onAppear {
+        if isAnimating {
+          isPulsing = true
+        }
+      }
+      .onChange(of: isAnimating) { _, animating in
+        if !animating { isPulsing = false }
+      }
+  }
+
+  private var isAnimating: Bool {
+    (bucket == "running" || bucket == "awaiting-input") && !reduceMotion
+  }
+}
+
+// MARK: - Type badge
+
+struct LaneTypeBadge: View {
+  let text: String
+  let tint: Color
+
+  var body: some View {
+    Text(text)
+      .font(.caption2.weight(.semibold))
+      .foregroundStyle(tint)
+      .padding(.horizontal, 7)
+      .padding(.vertical, 3)
+      .background(tint.opacity(0.12), in: Capsule())
+      .glassEffect()
+  }
+}
+
+// MARK: - Micro chip
+
+struct LaneMicroChip: View {
+  let icon: String
+  let text: String?
+  let tint: Color
+
+  var body: some View {
+    HStack(spacing: 3) {
+      Image(systemName: icon)
+        .font(.system(size: 8, weight: .semibold))
+      if let text {
+        Text(text)
+          .font(.system(.caption2).weight(.medium))
+      }
+    }
+    .foregroundStyle(tint)
+    .padding(.horizontal, 6)
+    .padding(.vertical, 3)
+    .background(tint.opacity(0.1), in: Capsule())
+    .glassEffect()
+  }
+}
+
+// MARK: - Action button
+
+struct LaneActionButton: View {
+  let title: String
+  let symbol: String
+  let tint: Color
+  let action: () -> Void
+
+  init(title: String, symbol: String, tint: Color = ADEColor.textSecondary, action: @escaping () -> Void) {
+    self.title = title
+    self.symbol = symbol
+    self.tint = tint
+    self.action = action
+  }
+
+  var body: some View {
+    Button(action: action) {
+      HStack(spacing: 5) {
+        Image(systemName: symbol)
+          .font(.system(size: 11, weight: .semibold))
+        Text(title)
+          .font(.caption.weight(.medium))
+      }
+      .foregroundStyle(tint)
+      .padding(.horizontal, 10)
+      .padding(.vertical, 7)
+      .background(tint.opacity(0.1), in: Capsule())
+      .glassEffect()
+    }
+    .buttonStyle(.plain)
+  }
+}
+
+// MARK: - Quick action
+
+struct LaneQuickAction: View {
+  let title: String
+  let symbol: String
+  let tint: Color
+  let action: () -> Void
+
+  var body: some View {
+    Button(action: action) {
+      VStack(spacing: 4) {
+        Image(systemName: symbol)
+          .font(.system(size: 16, weight: .medium))
+          .symbolRenderingMode(.hierarchical)
+        Text(title)
+          .font(.caption2.weight(.medium))
+      }
+      .foregroundStyle(tint)
+      .frame(width: 64, height: 54)
+      .background(ADEColor.surfaceBackground.opacity(0.08), in: RoundedRectangle(cornerRadius: 12, style: .continuous))
+      .glassEffect(in: .rect(cornerRadius: 12))
+      .overlay(
+        RoundedRectangle(cornerRadius: 12, style: .continuous)
+          .stroke(ADEColor.border.opacity(0.16), lineWidth: 0.5)
+      )
+    }
+    .buttonStyle(ADEScaleButtonStyle())
+  }
+}
+
+// MARK: - Menu label
+
+struct LaneMenuLabel: View {
+  let title: String
+
+  var body: some View {
+    HStack(spacing: 4) {
+      Text(title)
+        .font(.caption.weight(.medium))
+      Image(systemName: "chevron.down")
+        .font(.system(size: 8, weight: .bold))
+    }
+    .foregroundStyle(ADEColor.textSecondary)
+    .padding(.horizontal, 10)
+    .padding(.vertical, 7)
+    .background(ADEColor.surfaceBackground.opacity(0.55), in: Capsule())
+    .glassEffect()
+    .overlay(
+      Capsule()
+        .stroke(ADEColor.border.opacity(0.16), lineWidth: 0.5)
+    )
+  }
+}
+
+// MARK: - Open chip
+
+struct LaneOpenChip: View {
+  let snapshot: LaneListSnapshot
+  let isPinned: Bool
+
+  var body: some View {
+    HStack(spacing: 6) {
+      Circle()
+        .fill(runtimeTint(bucket: snapshot.runtime.bucket))
+        .frame(width: 6, height: 6)
+      Text(snapshot.lane.name)
+        .font(.caption.weight(.medium))
+        .foregroundStyle(ADEColor.textPrimary)
+        .lineLimit(1)
+      if isPinned {
+        Image(systemName: "pin.fill")
+          .font(.system(size: 8))
+          .foregroundStyle(ADEColor.accent)
+      }
+    }
+    .padding(.horizontal, 10)
+    .padding(.vertical, 7)
+    .background(ADEColor.surfaceBackground.opacity(0.55), in: Capsule())
+    .glassEffect()
+    .overlay(
+      Capsule()
+        .stroke(ADEColor.border.opacity(0.16), lineWidth: 0.5)
+    )
+    .accessibilityLabel("\(snapshot.lane.name)\(isPinned ? ", pinned" : "")")
+  }
+}
+
+// MARK: - Launch tile
+
+struct LaneLaunchTile: View {
+  let title: String
+  let symbol: String
+  let tint: Color
+  let action: () -> Void
+
+  var body: some View {
+    Button(action: action) {
+      VStack(spacing: 8) {
+        Image(systemName: symbol)
+          .font(.system(size: 18, weight: .semibold))
+          .symbolRenderingMode(.hierarchical)
+        Text(title)
+          .font(.caption.weight(.medium))
+      }
+      .foregroundStyle(tint)
+      .frame(maxWidth: .infinity)
+      .padding(.vertical, 14)
+      .background(ADEColor.surfaceBackground.opacity(0.08), in: RoundedRectangle(cornerRadius: 12, style: .continuous))
+      .glassEffect(in: .rect(cornerRadius: 12))
+      .overlay(
+        RoundedRectangle(cornerRadius: 12, style: .continuous)
+          .stroke(tint.opacity(0.14), lineWidth: 0.5)
+      )
+    }
+    .buttonStyle(ADEScaleButtonStyle())
+    .accessibilityLabel("Launch \(title)")
+  }
+}
+
+// MARK: - Session card
+
+struct LaneSessionCard: View {
+  let session: TerminalSessionSummary
+
+  var body: some View {
+    VStack(alignment: .leading, spacing: 6) {
+      HStack {
+        Text(session.title)
+          .font(.subheadline.weight(.semibold))
+          .foregroundStyle(ADEColor.textPrimary)
+        Spacer()
+        LaneTypeBadge(text: session.status.uppercased(), tint: session.status == "running" ? ADEColor.success : ADEColor.textSecondary)
+      }
+      if let preview = session.lastOutputPreview {
+        Text(preview)
+          .font(.caption)
+          .foregroundStyle(ADEColor.textMuted)
+          .lineLimit(2)
+      }
+    }
+    .adeGlassCard(cornerRadius: 10, padding: 10)
+  }
+}
+
+// MARK: - Chat card
+
+struct LaneChatCard: View {
+  let chat: AgentChatSessionSummary
+
+  var body: some View {
+    VStack(alignment: .leading, spacing: 6) {
+      HStack {
+        Text(chat.title ?? chat.provider.uppercased())
+          .font(.subheadline.weight(.semibold))
+          .foregroundStyle(ADEColor.textPrimary)
+        Spacer()
+        LaneTypeBadge(text: chat.status.uppercased(), tint: chat.status == "active" ? ADEColor.success : ADEColor.textSecondary)
+      }
+      Text(chat.model)
+        .font(.system(.caption, design: .monospaced))
+        .foregroundStyle(ADEColor.textSecondary)
+      if let preview = chat.lastOutputPreview {
+        Text(preview)
+          .font(.caption)
+          .foregroundStyle(ADEColor.textMuted)
+          .lineLimit(2)
+      }
+    }
+    .adeGlassCard(cornerRadius: 10, padding: 10)
+  }
+}
+
+// MARK: - Info row
+
+struct LaneInfoRow: View {
+  let label: String
+  let value: String
+  var isMonospaced = false
+
+  var body: some View {
+    HStack(alignment: .top, spacing: 12) {
+      Text(label)
+        .font(.caption.weight(.semibold))
+        .foregroundStyle(ADEColor.textMuted)
+        .frame(width: 54, alignment: .leading)
+      Text(value)
+        .font(isMonospaced ? .system(.caption, design: .monospaced) : .subheadline)
+        .foregroundStyle(ADEColor.textPrimary)
+        .frame(maxWidth: .infinity, alignment: .leading)
+    }
+  }
+}
+
+// MARK: - Text field
+
+struct LaneTextField: View {
+  let title: String
+  @Binding var text: String
+
+  init(_ title: String, text: Binding<String>) {
+    self.title = title
+    self._text = text
+  }
+
+  var body: some View {
+    TextField(title, text: $text, axis: .vertical)
+      .textFieldStyle(.plain)
+      .foregroundStyle(ADEColor.textPrimary)
+      .adeInsetField()
+  }
+}
+
+// MARK: - Scale button style
+
+struct ADEScaleButtonStyle: ButtonStyle {
+  func makeBody(configuration: Configuration) -> some View {
+    configuration.label
+      .scaleEffect(configuration.isPressed ? 0.97 : 1.0)
+      .opacity(configuration.isPressed ? 0.85 : 1.0)
+      .animation(.snappy(duration: 0.2), value: configuration.isPressed)
+  }
+}
+
+// MARK: - Lane list row
+
+struct LaneListRow: View {
+  let snapshot: LaneListSnapshot
+  let isPinned: Bool
+  let isOpen: Bool
+
+  var body: some View {
+    HStack(spacing: 14) {
+      LaneStatusIndicator(bucket: snapshot.runtime.bucket)
+      VStack(alignment: .leading, spacing: 5) {
+        HStack(spacing: 8) {
+          Text(snapshot.lane.name)
+            .font(.body.weight(.semibold))
+            .foregroundStyle(ADEColor.textPrimary)
+            .lineLimit(1)
+          if snapshot.lane.laneType == "primary" {
+            LaneTypeBadge(text: "Primary", tint: ADEColor.accent)
+          } else if snapshot.lane.laneType == "attached" {
+            LaneTypeBadge(text: "Attached", tint: ADEColor.textMuted)
+          }
+        }
+        Text(snapshot.lane.branchRef)
+          .font(.system(.caption, design: .monospaced))
+          .foregroundStyle(ADEColor.textSecondary)
+          .lineLimit(1)
+        HStack(spacing: 6) {
+          if snapshot.lane.status.ahead > 0 {
+            LaneMicroChip(icon: "arrow.up", text: "\(snapshot.lane.status.ahead)", tint: ADEColor.success)
+          }
+          if snapshot.lane.status.behind > 0 {
+            LaneMicroChip(icon: "arrow.down", text: "\(snapshot.lane.status.behind)", tint: ADEColor.warning)
+          }
+          if snapshot.runtime.sessionCount > 0 {
+            LaneMicroChip(
+              icon: runtimeSymbol(snapshot.runtime.bucket),
+              text: "\(snapshot.runtime.sessionCount)",
+              tint: runtimeTint(bucket: snapshot.runtime.bucket)
+            )
+          }
+          if snapshot.lane.childCount > 0 {
+            LaneMicroChip(icon: "square.stack.3d.up", text: "\(snapshot.lane.childCount)", tint: ADEColor.textMuted)
+          }
+          if isPinned {
+            Image(systemName: "pin.fill")
+              .font(.system(size: 9))
+              .foregroundStyle(ADEColor.accent)
+          }
+        }
+        if let activity = laneActivitySummary(snapshot) {
+          Text(activity)
+            .font(.caption2)
+            .foregroundStyle(ADEColor.textMuted)
+            .lineLimit(1)
+        }
+      }
+      Spacer(minLength: 8)
+      VStack(alignment: .trailing, spacing: 6) {
+        lanePriorityBadge(snapshot: snapshot)
+      }
+      Image(systemName: "chevron.right")
+        .font(.caption2.weight(.semibold))
+        .foregroundStyle(ADEColor.textMuted)
+    }
+    .adeGlassCard(cornerRadius: 16, padding: 14)
+    .overlay(
+      RoundedRectangle(cornerRadius: 16, style: .continuous)
+        .stroke(isOpen ? ADEColor.accent.opacity(0.35) : ADEColor.border.opacity(0.14), lineWidth: isOpen ? 1 : 0.75)
+    )
+    .accessibilityElement(children: .combine)
+    .accessibilityLabel(laneRowAccessibilityLabel)
+  }
+
+  private var laneRowAccessibilityLabel: String {
+    var parts = [snapshot.lane.name, snapshot.lane.branchRef]
+    if snapshot.lane.status.dirty { parts.append("dirty") }
+    if isPinned { parts.append("pinned") }
+    if isOpen { parts.append("open") }
+    if snapshot.lane.status.ahead > 0 { parts.append("\(snapshot.lane.status.ahead) ahead") }
+    if snapshot.lane.status.behind > 0 { parts.append("\(snapshot.lane.status.behind) behind") }
+    return parts.joined(separator: ", ")
+  }
+}
